chore: 패키지 이동
exception, config -> common
