+++ src/main/java/in/koreatech/payment/common/exception/custom/KoinIllegalArgumentException.java
@@ -1,4 +1,4 @@
-package in.koreatech.payment.exception.custom;
+package in.koreatech.payment.common.exception.custom;
 
 public class KoinIllegalArgumentException extends KoinException {
