+++ src/main/java/in/koreatech/payment/common/exception/custom/KoinIllegalStateException.java
@@ -1,4 +1,4 @@
-package in.koreatech.payment.exception.custom;
+package in.koreatech.payment.common.exception.custom;
 
 public class KoinIllegalStateException extends KoinException {
